refactor(ui): align context and route constants structure
