+++ src/ui/components/organisms/ProductFilters.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from "vitest";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { initialFilterState } from "@/ui/stores/context/productFilter";
+import { initialFilterState } from "@/ui/context/productFilter";
 import type { Product } from "@/core/domain";
 import { ProductFilters } from "./ProductFilters";
