+++ src/ui/hooks/AGENTS.md
@@ -1,6 +1,6 @@
 # AGENTS.md — src/ui/hooks/
 
-> Last updated: 2026-07-28
+> Last updated: 2026-08-26
 
 ## Overview
 
@@ -20,7 +20,7 @@ src/ui/hooks/
 
 - 파일명은 camelCase, `use` 접두사 필수 — 데이터 페칭 훅은 도메인을 PascalCase로 이어붙이고(`useCoupleInfo.ts`), 페칭 외 공유 로직 훅은 목적을 PascalCase로 이어붙인다(`useCountdown.ts`).
 - **"use client" 지시어를 파일 최상단에 고정한다 — 예외 없다.** 훅은 정의상 React hook을 호출하므로 client 전용이다. 호출하는 쪽 컴포넌트가 이미 client 경계 안이라서 괜찮겠거니 하고 생략하지 않는다 — `index.ts` 배럴이 형제 파일 전체를 하나의 모듈 그래프로 묶기 때문에, 이 파일 하나가 경계 선언 없이 hook을 쓰면 그 배럴을 조금이라도 참조하는 아무 Server Component에서나 빌드 에러가 난다(Gotchas 참고).
-- 외부 SDK를 초기화하는 훅(`useKakaoLoader` 등)을 여기 두지 않는다 — 훅이라도 연동 대상 하나에 강결합돼 있으면 `src/ui/lib/{service}/`의 "연동 대상 1개 = 폴더 1개" 소관이다(`src/ui/lib/AGENTS.md` 참고).
+- 외부 SDK를 초기화하는 훅(`useKakaoLoader` 등)을 여기 두지 않는다 — 훅이라도 연동 대상 하나에 강결합돼 있으면 `src/adapters/browser/{service}/`의 "연동 대상 1개 = 폴더 1개" 소관이다(`src/adapters/AGENTS.md` 참고).
 
 ## Gotchas
 
@@ -28,4 +28,4 @@ src/ui/hooks/
 
 ## 관련 문서
 
-- 외부 SDK 초기화 훅의 실제 위치(lib과의 배치 경계): `src/ui/lib/AGENTS.md`
+- 외부 SDK 초기화 훅의 실제 위치(adapters와의 배치 경계): `src/adapters/AGENTS.md`
